changed name for the taskContext and moved add reward to scoreboard to it
