+++ src/components/ToDos/TasksCategoryPage.tsx
@@ -1,6 +1,6 @@
 import { StyleSheet, Text, View, Image } from 'react-native';
 import React, { ReactElement, useEffect, useState } from 'react';
-import { useTasks } from '../../util/context/TaskContext';
+import { useTasks } from '../../util/context/AddtoDBContext';
 import { Tasks } from '../../Interfaces';
 import TaskCard from './TaskCard';
 import { useDimensions } from '@react-native-community/hooks';
@@ -147,3 +147,4 @@ export const TasksCategoryPage = ({ category }: TasksCategoryPageProps) => {
     </View>
   );
 };
+
